Lock contention
Lock contention takes place when a thread tries to acquire the lock to
an object which is already acquired by other thread. Until the object is
released, the thread is blocked (in other words, it is in the Waiting
state). In some cases, this may lead to a so-called serial execution
which negatively affects application performance. The serial execution
pattern can be easily determined using the Threads diagram. Instead of
threads running in parallel, you will see only one thread running at a
time.